requirements.txt
notionai_py>=0.1.4
click>=8.1.3
pynput>=1.7.6
pyperclip>=1.8.2
fastavro==1.7.0
pyllms
# openai
# tiktoken
# anthropic==0.3
# ai21
# cohere
# aleph-alpha-client
# huggingface_hub
# prettytable
# aiohttp
# google-cloud-aiplatform~=1.28.1
# einops
# accelerate
# protobuf~=3.2.3
# grpcio~=1.54.2